cpu: Added correct return type for ROB::countInsts
- return size_t (unsigned) according to the .size() return type
- fixed typo in doc (source of warning with some compilers)

Reviewed-on: https://gem5-review.googlesource.com/c/public/gem5/+/18948
Maintainer: Jason Lowe-Power <jason@lowepower.com>
Tested-by: kokoro <noreply+kokoro@google.com>

+++ src/cpu/o3/rob.hh
@@ -258,7 +258,7 @@ class ROB
      *  threadEntries to get the instructions in the ROB unless you are
      *  double checking that variable.
      */
-    int countInsts(ThreadID tid);
+    size_t countInsts(ThreadID tid);
 
     /** Registers statistics. */
     void regStats();
